Supports minification using angular.module right.
Using "def" and extend.js my custom namespace scheme worked fine when
the output was not minified, but broke due to angulars dependency 
injection method not being able to reference minified argument names to
controllers and other functions.

It seems the best way around this is to forego using some custom 
namespacing scheme and just rely on angular.module() and being explicit
about dependencies by defining them as strings before defining the 
module/factory/controller/directive you are creating.

There is probably some magic that could be done by modifying "def" so
that it handled the dependencies for you but I'm not sure if it's worth
the hassle to break from angular conventions.